Implemented font and UI

+++ ZombieGame/Pathfinder.h
@@ -19,17 +19,17 @@ public:
 	void setTileWidth(int tileWidth) { m_tileWidth = tileWidth; }
 
 private:	
-	glm::vec2 convertPositionToCoordinates(glm::vec2 position);
+	glm::ivec2 convertPositionToCoordinates(glm::vec2 position);
 
-	void updateNeighbourNodes(glm::vec2 nodeCoords, glm::vec2 targetCoords);
-	void updateNode(float baseNodeX, float baseNodeY, float updateNodeX, float updateNodeY, glm::vec2 targetCoords, int addedG);
-	glm::vec2 getLowestF();
+	void updateNeighbourNodes(glm::ivec2 nodeCoords, glm::ivec2 targetCoords);
+	void updateNode(int baseNodeX, int baseNodeY, int updateNodeX, int updateNodeY, glm::ivec2 targetCoords, int addedG);
+	glm::ivec2 getLowestF();
 
-	glm::vec2 m_previousTargetCoords = { 0, 0 };
-	glm::vec2 m_previousStartCoords = { 0, 0 };
+	glm::ivec2 m_previousTargetCoords = { 0, 0 };
+	glm::ivec2 m_previousStartCoords = { 0, 0 };
 	std::vector<std::vector<Node>> m_field;
-	std::vector<glm::vec2> m_openList;
-	std::vector<glm::vec2> m_closedList;
+	std::vector<glm::ivec2> m_openList;
+	std::vector<glm::ivec2> m_closedList;
 
 	int m_tileWidth;
 };
